backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.7.0
python-multipart>=0.0.9
google-genai>=1.0.0.0
pymupdf>=1.24.0
sqlalchemy>=2.0.29
pinecone>=5.0.0
python-dotenv>=1.0.1
rapidocr-onnxruntime>=1.2.3
pillow>=12.0.0
